{"schema":"decodex/commit/1","summary":"Split progressive search type modules","authority":"manual"}
